requirements.txt
apache-beam[gcp,test]==2.29.0
google-cloud-storage==1.38.0
